fix(server): migration drop useless collection if exists

+++ server/migrations/20230206153259-drop-useless-collections.js
@@ -1,11 +1,15 @@
 export const up = async (db) => {
-  // Suppression des collections inutilisées
-  await db.collection("effectifsApprenants").drop();
-  await db.collection("referentielSiret").drop();
-  await db.collection("duplicatesEvents").drop();
-  await db.collection("cfas").drop();
-  await db.collection("archiveDossiersApprenants").drop();
-  await db.collection("dossiersApprenants").drop();
+  // Suppression des collections inutilisées (si elles existent)
+  const collectionsInDb = await db.listCollections().toArray();
+
+  if (existInDb(collectionsInDb, "effectifsApprenants")) await db.collection("effectifsApprenants").drop();
+  if (existInDb(collectionsInDb, "referentielSiret")) await db.collection("referentielSiret").drop();
+  if (existInDb(collectionsInDb, "duplicatesEvents")) await db.collection("duplicatesEvents").drop();
+  if (existInDb(collectionsInDb, "cfas")) await db.collection("cfas").drop();
+  if (existInDb(collectionsInDb, "archiveDossiersApprenants")) await db.collection("archiveDossiersApprenants").drop();
+  if (existInDb(collectionsInDb, "dossiersApprenants")) await db.collection("dossiersApprenants").drop();
 };
+
+const existInDb = (collectionsInDb, collectionName) => collectionsInDb.map(({ name }) => name).includes(collectionName);
 
 export const down = async () => {};
